refactoring Merged the db_methods package with queries package. Now, we will have only one package for all db interactions. Made the queries_api DB methods more testable.
