+++ internal/endpoints/statuses.go
@@ -7,13 +7,13 @@ import (
 	"strings"
 	"time"
 
-	"github.com/redhatinsights/payload-tracker-go/internal/db_methods"
+	"github.com/redhatinsights/payload-tracker-go/internal/queries"
 	l "github.com/redhatinsights/payload-tracker-go/internal/logging"
 	"github.com/redhatinsights/payload-tracker-go/internal/structs"
 )
 
 var (
-	RetrieveStatuses = db_methods.RetrieveStatuses
+	RetrieveStatuses = queries.RetrieveStatuses
 )
 
 func Statuses(w http.ResponseWriter, r *http.Request) {
@@ -42,7 +42,7 @@ func Statuses(w http.ResponseWriter, r *http.Request) {
 		writeResponse(w, http.StatusBadRequest, getErrorBody(message, http.StatusBadRequest))
 		return
 	}
-	count, payloads := RetrieveStatuses(q)
+	count, payloads := RetrieveStatuses(dbObj, q)
 	duration := time.Since(start).Seconds()
 
 	statusesData := structs.StatusesData{count, duration, payloads}
